Fix gb/lb remove not showing message, update help of /fly and fix it to actually work oops

+++ MCGalaxy/Network/Utils/NetUtils.cs
@@ -69,9 +69,9 @@ namespace MCGalaxy {
                 WriteI16((short)pos.Y, arr, offset + 2);
                 WriteI16((short)pos.Z, arr, offset + 4);
             } else {
-                WriteI32((int)pos.X, arr, offset + 0);
-                WriteI32((int)pos.Y, arr, offset + 4);
-                WriteI32((int)pos.Z, arr, offset + 8);
+                WriteI32(pos.X, arr, offset + 0);
+                WriteI32(pos.Y, arr, offset + 4);
+                WriteI32(pos.Z, arr, offset + 8);
             }
             return extPos ? 12 : 6;
         }
